refactor: test folder restructure
